refactor(sparkeks): move entrypoint strategy to entrypoint.go; drop added tests
Move entrypointStrategy + jar/sqlWrapper strategies + newEntrypointStrategy and
jarApplicationType into internal/pkg/object/command/sparkeks/entrypoint.go.
sparkeks.go keeps applySparkOperatorConfig (delegates to newEntrypointStrategy)
and the Gap-4 default. Revert the added sparkeks_test.go unit tests per request.
Build + vet clean.
